fix: acronyms not converted correctly to Pascal case
Some Pascal case strings are not converted correctly, where double
uppercase letters were being removed.
I had just installed this module and Jellyfin from scratch and noticed
weird things on my `network.xml`.
The `enableIPv4` option ended up as `EnableP4`, `enableUPnP` was broken
in a similar fashion.
I have not tested this extensively but it fixes my case, and the few
tests there are in there are passing (I added a couple as well).
It is a very naive fix, though.